% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autotracer_linelist}
\alias{autotracer_linelist}
\title{Simulated Data for Illustration
\format{a data.frame with 10,5000 rows and 9 columns}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10500 rows and 9 columns.
}
\usage{
autotracer_linelist
}
\description{
Simulated Data for Illustration
\format{a data.frame with 10,5000 rows and 9 columns}
}
\keyword{datasets}
