% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_probable_cases.R
\name{connect_probable_cases}
\alias{connect_probable_cases}
\title{Connect Probable Cases}
\usage{
connect_probable_cases(
  dat,
  weights_in = NULL,
  threshold = 30,
  exposure_link = NULL
)
}
\arguments{
\item{dat}{the dataframe of likely connect cases with a column named "date"
indicating the onset or positive test date and "patient_id", a unique
identifier for the record.}

\item{weights_in}{the weights to use for the serial interval if available}

\item{threshold}{integer, the threshold in days at which to discard a connected
case (e.g. >30 days from previous case, then discard).}

\item{exposure_link}{string or column name, how these cases are connected}
}
\description{
Probabilistically estimates a likely tranmission chain using EMR
derrived data.
}
