% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autotracer_linelist}
\alias{autotracer_linelist}
\title{Simulated Patient Data
Synthetic data that can be used to explore "autotracer"}
\format{
a data.frame with 10500 rows and 9 columns:
\describe{
\item{x}{Latitude}
\item{y}{Longitude}
\item{Race}{Race indentifier}
\item{patient_id}{Unique patient indentifier (e.g. MRN)}
\item{Age}{Age of the patient in years}
\item{Sex}{Biological sex of the patient}
\item{Language}{Primary spoken language of the patient}
\item{Age10}{Age given as a ten year age group}
\item{date}{Date of test or infection identification}
}
}
\usage{
autotracer_linelist
}
\description{
Simulated Patient Data
Synthetic data that can be used to explore "autotracer"
}
\keyword{datasets}
